Add an inspection for expensive fields
