+++ intellij-plugin/src/main/resources/inspectionDescriptions/ApolloFieldInsights.html
@@ -1,0 +1,12 @@
+<html>
+<body>
+Reports fields that are particularly expensive to resolve.
+<p>
+    Field metrics are gathered from Apollo GraphOS.<br>
+    See the <a href="https://www.apollographql.com/docs/graphos/metrics/field-usage">GraphOS documentation</a> on how to enable and use this feature.
+</p>
+<!-- tooltip end -->
+<p>To retrieve the data, configure your GraphOS API keys in <a href="settings://com.apollographql.ijplugin.settings.SettingsConfigurable">Languages&nbsp;&amp;&nbsp;Frameworks | GraphQL | Apollo&nbsp;Kotlin</a>
+</p>
+</body>
+</html>
