+++ intellij-plugin/src/main/resources/messages/ApolloBundle.properties
@@ -55,5 +55,8 @@ CompatToOperationBasedCodegenMigrationProcessor.noUsage=No compat codegen usage 
 
 SandboxService.OpenInSandboxAction.text=Open in Apollo Sandbox
 
-FieldInsightsInlayHintsProvider.settings.name=Apollo Field Insights
-FieldInsightsInlayHintsProvider.settings.description=Show field latencies gathered within Apollo GraphOS.<br><br><a href="https://www.apollographql.com/docs/graphos/metrics/field-usage">Learn more.</a><br><br>Configure your GraphOS API keys in <kbd>Languages\\&nbsp;\\&amp;\\&nbsp;Frameworks</kbd> | <kbd>GraphQL</kbd> | <kbd>Apollo\\&nbsp;Kotlin</kbd>.
+inspection.group.graphql.studio=Apollo GraphOS
+inspection.group.graphql=GraphQL
+inspection.fieldInsights.displayName=Expensive field
+inspection.fieldInsights.reportText=<html><b><tt>{0}</tt></b> is expensive to resolve: <b>{1}</b>
+inspection.fieldInsights.settings.threshold=Threshold (ms)
